SDL_GPUColorComponentFlags

+++ Engine/Internals/SDL/Enums/SDL_GPUColorComponentFlags.cs
@@ -5,11 +5,11 @@ using System;
 namespace Engine.SDL3
 {
     [Flags]
-    internal enum SDL_GPUColorComponentFlags : uint
+    internal enum SDL_GPUColorComponentFlags : byte
     {
-	    SDL_GPU_COLORCOMPONENT_R = (1U << 0),
-	    SDL_GPU_COLORCOMPONENT_G = (1U << 1),
-	    SDL_GPU_COLORCOMPONENT_B = (1U << 2),
-	    SDL_GPU_COLORCOMPONENT_A = (1U << 3),
+	    SDL_GPU_COLORCOMPONENT_R = (1 << 0),
+	    SDL_GPU_COLORCOMPONENT_G = (1 << 1),
+	    SDL_GPU_COLORCOMPONENT_B = (1 << 2),
+	    SDL_GPU_COLORCOMPONENT_A = (1 << 3),
     }
 }
